tidied up step 2 page and made some features to clean
