+++ app/(modals)/create/post/step2/_features/useTimePicker.ts
@@ -1,0 +1,17 @@
+import { format } from "date-fns";
+import { useState } from "react";
+
+export const useTimePicker = (today: Date) => {
+  const [time, setTime] = useState<Date>(today);
+  const [showTimePicker, setShowTimePicker] = useState<boolean>(false);
+
+  const formattedTime = format(time, "HH:mm");
+
+  return {
+    time,
+    setTime,
+    showTimePicker,
+    setShowTimePicker,
+    formattedTime,
+  };
+};
